feat!: Rework entire movement system as a whole.
The movement system was prior based on instructions to move. Now, it is
based on coordinate input. The PI will send a coordinate that
is mapped as the steps in x and y as steps, and the
movement system will parse and execute on that.

Further, much logic originally in Patchy.ino has been moved to
PatchySystems.* to more accurately map behavior to the entire object
that is Patchy.

As consequence, many core changes were made.